Add's IPv6 support.
IPv6 connections needs to use AF_INET6 instead of AF_INET so a function called "is_it_ipv6" was added to test the host to see if IPv4 or IPv6 was used and then uses a AF_INET6 instead if IPv6 is detected in fetch_data. Also entry2url needed to be modified as a IPv6 address like gopher://[::1]/ would turn into gopher://::1/ which is not accessible.